pyproject.toml
[tool.poetry]
name = "paytaca-gifts"
version = "0.1.0"
description = ""
authors = ["Marie Edcel Balite <mdrbalite@gmail.com>"]
readme = "README.md"
packages = [{include = "paytaca_gifts"}]

[tool.poetry.dependencies]
python = "^3.8"
sanic = "^22.6.2"
tortoise-orm = "^0.19.2"
aerich = "^0.6.3"
celery = "^5.2.7"
redis = "^4.3.4"
environs = "^9.5.0"
sanic-testing = "^22.6.0"
pytest-asyncio = "^0.19.0"
sanic-ext = "^22.6.3"
asyncpg = "^0.26.0"
supervisor = "^4.2.4"
sanic-validation = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
